requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
django-redis>=5.4
django-filter>=23.0
django-import-export>=3.3
gunicorn>=21.2
whitenoise>=6.5
python-dotenv>=1.0
